Delete memoized value in a strong and stable order.
This commit deletes records from the db_memoized table in a predefined 
and stable order. This should make sure that two transactions that 
should delete the same records, both do this in the same order, 
preventing dead lock situations.

Background: postgres does not let a user define the order in which rows
are deleted from a table, yet to have two transactions able to delete
the same row and not to deadlock we need something like this (or we
need to employ some locking mechanism instead.)

This patch runs a loop on the database server to delete all records
one by one, in ctid order. ctid ordering is stable unless one does
a DDL statement or VACUUMs the table, so I think we should be much
better than before.